# Speedup InceptionV3 and GoogleNet tests on windows (upgrade scipy)