--- a/.circleci/unittest/windows/scripts/environment.yml
+++ b/.circleci/unittest/windows/scripts/environment.yml
@@ -13,6 +13,6 @@ dependencies:
   - pip:
     - future
     - pillow>=4.1.1
-    - scipy==1.4.1
+    - scipy
     - av
     - dataclasses