Rough improvement: ~96%
- Before: 931.24s - [reference](https://app.circleci.com/pipelines/github/pytorch/vision/5594/workflows/a84d80ba-a832-4ca6-856d-fde31313bc24/jobs/353553)
```
327.29s call     test/test_models.py::ModelTester::test_inception_v3_cpu
326.23s call     test/test_models.py::ModelTester::test_inception_v3_cuda
135.34s call     test/test_models.py::ModelTester::test_googlenet_cuda
133.91s call     test/test_models.py::ModelTester::test_googlenet_cpu
5.43s call     test/test_models.py::ModelTester::test_inceptionv3_eval
3.04s call     test/test_models.py::ModelTester::test_googlenet_eval
```
- After: 32.38s - [reference](https://app.circleci.com/pipelines/github/pytorch/vision/5595/workflows/85e9e739-c5ec-4704-9854-8e2803a37257/jobs/353561)
```
7.97s call     test/test_models.py::ModelTester::test_googlenet_cuda
7.41s call     test/test_models.py::ModelTester::test_inception_v3_cpu
5.24s call     test/test_models.py::ModelTester::test_inception_v3_cuda
4.52s call     test/test_models.py::ModelTester::test_googlenet_cpu
4.50s call     test/test_models.py::ModelTester::test_inception_v3_eval
2.74s call     test/test_models.py::ModelTester::test_googlenet_eval
```

The above are rough estimations based on a couple or runs.

The problem is traced to the use of scipy in the weight initialization of the two models:
https://github.com/pytorch/vision/blob/90645ccd0e774ad76200245e32222a23d09f2312/torchvision/models/inception.py#L122-L124

https://github.com/pytorch/vision/blob/90645ccd0e774ad76200245e32222a23d09f2312/torchvision/models/googlenet.py#L126-L128

The slow down is caused by a regression on scipy affecting versions 1.4.x: https://github.com/scipy/scipy/issues/11299. The reason why only Windows runs are affected is because our CI config is pinning the version of scipy to 1.4.1 on Windows but not on Linux. This was introduced at #2328 (first commit of the file) and it's probably by accident that the two files are not aligned. 

By unpinning the version and allowing scipy to use newer versions on Windows, the issue is fixed. This is an alternative solution to #3195.